// src/context/questionnaire/enum/questionnaireResultType.enum.ts
export enum QuestionnaireResultType {
  OnmRmRisk = 'onm-rmRisk',
  OnmRmSuspectionOnStageZero = 'onm-rmSuspectionOnStageZero',
  OnmRmEstablished = 'onm-rmEstablished',
  OnmRmInsignificantRisk = 'onm-rmInsignificantRisk',
  StageOne = 'stageOne',
  StageTwo = 'stageTwo',
  StageThree = 'stageThree',
}